- extract test enums to separate object - split testing into reads and writes specs
